mrthreshold: Fix signed - unsigned comparisons

+++ cmd/mrthreshold.cpp
@@ -236,7 +236,7 @@ default_type calculate (Image<value_type>& in,
     const ssize_t index (bottom >= 0 ?
                          size_t(bottom) - 1 :
                          (ssize_t(data.size()) - ssize_t(top)));
-    if (index < 0 || index >= data.size())
+    if (index < 0 || index >= ssize_t(data.size()))
       throw Exception ("Number of valid input image values (" + str(data.size()) + ") less than number of voxels requested via -" + (bottom >= 0 ? "bottom" : "top") + " option (" + str(index) + ")");
     std::nth_element (data.begin(), data.begin() + index, data.end());
     const value_type threshold_float = data[index];
@@ -245,7 +245,7 @@ default_type calculate (Image<value_type>& in,
       if (data[index-1] == threshold_float)
         issue_degeneracy_warning = true;
     }
-    if (index < data.size() - 1) {
+    if (index < ssize_t(data.size()) - 1) {
       std::nth_element (data.begin(), data.begin() + index + 1, data.end());
       if (data[index+1] == threshold_float)
         issue_degeneracy_warning = true;
